+++ nfapi/open-nFAPI/vnf/public_inc/nfapi_vnf_interface.h
@@ -33,7 +33,7 @@ typedef struct nfapi_vnf_phy_info
 	int phy_id; //phy_id
 
 	/*! Timing window */
-	uint8_t timing_window;
+	uint16_t timing_window;
 	/*! Timing info mode */
 	uint8_t timing_info_mode;
 	/*! Timing info period */
